--- src/com/zilen/messagingService/entity/attachment/Attachment.java
@@ -1,4 +1,0 @@
-package com.zilen.messagingService.entity.attachment;
-
-public interface Attachment {
-}
