--- src/com/zilen/messagingService/entity/channel/Channel.java
@@ -1,6 +1,0 @@
-package com.zilen.messagingService.entity.channel;
-
-public interface Channel {
-
-    String getUserName();
-}
