--- src/com/zilen/messagingService/entity/channel/Facebook.java
@@ -1,18 +1,0 @@
-package com.zilen.messagingService.entity.channel;
-
-import lombok.Builder;
-import lombok.Data;
-import lombok.NonNull;
-
-import java.util.UUID;
-
-@Data
-@Builder(toBuilder = true)
-public class Facebook implements Channel {
-
-    @NonNull
-    private UUID id;
-    private String userName;
-    private String fromFacebookId;
-    private String toFacebookId;
-}
