+++ src/main/java/com/zilen/messagingservice/entity/Message.java
@@ -1,0 +1,26 @@
+package com.zilen.messagingservice.entity;
+
+import com.zilen.messagingservice.entity.attachment.Attachment;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NonNull;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.UUID;
+
+@Data
+@Builder(toBuilder = true)
+public class Message {
+
+    @NonNull
+    private UUID id;
+    @NonNull
+    private String userName;
+    @NonNull
+    private String text;
+    @NonNull
+    private LocalDateTime dateTime;
+    private List<Attachment> attachments;
+
+}
