+++ src/main/java/com/zilen/messagingservice/entity/attachment/Attachment.java
@@ -1,0 +1,4 @@
+package com.zilen.messagingservice.entity.attachment;
+
+public interface Attachment {
+}
