+++ src/main/java/com/zilen/messagingservice/entity/attachment/Document.java
@@ -1,0 +1,19 @@
+package com.zilen.messagingservice.entity.attachment;
+
+import lombok.Builder;
+import lombok.Data;
+import lombok.NonNull;
+
+import java.util.UUID;
+
+@Data
+@Builder(toBuilder = true)
+public class Document implements Attachment {
+
+    @NonNull
+    private UUID id;
+    @NonNull
+    private String title;
+    private String format;
+    private int size;
+}
