+++ src/main/java/com/zilen/messagingservice/entity/channel/Channel.java
@@ -1,0 +1,6 @@
+package com.zilen.messagingservice.entity.channel;
+
+public interface Channel {
+
+    String getUserName();
+}
