+++ src/main/java/com/zilen/messagingservice/entity/channel/Email.java
@@ -1,0 +1,18 @@
+package com.zilen.messagingservice.entity.channel;
+
+import lombok.Builder;
+import lombok.Data;
+import lombok.NonNull;
+
+import java.util.UUID;
+
+@Data
+@Builder(toBuilder = true)
+public class Email implements Channel {
+
+    @NonNull
+    private UUID id;
+    private String userName;
+    private String fromEmailAddress;
+    private String toEmailAddress;
+}
