+++ src/test/java/com/zilen/messagingservice/MessagingServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.zilen.messagingservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MessagingServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
